+++ exonum-java-binding/testkit/src/test/java/com/exonum/binding/testkit/TestKitTestUtils.java
@@ -34,7 +34,7 @@ final class TestKitTestUtils {
   static final String ARTIFACT_FILENAME = "test-service.jar";
   private static final String ARTIFACT_VERSION = "1.0.0";
   static final ServiceArtifactId ARTIFACT_ID =
-      ServiceArtifactId.newJavaId("com.exonum.binding:test-service:" + ARTIFACT_VERSION);
+      ServiceArtifactId.newJavaId("com.exonum.binding/test-service", ARTIFACT_VERSION);
   static final String SERVICE_NAME = "test-service";
   static final int SERVICE_ID = 46;
   static final String CONFIGURATION_VALUE = "Initial value";
@@ -45,7 +45,7 @@ final class TestKitTestUtils {
   static final String ARTIFACT_FILENAME_2 = "test-service-2.jar";
   private static final String ARTIFACT_VERSION_2 = "2.8.0";
   static final ServiceArtifactId ARTIFACT_ID_2 =
-      ServiceArtifactId.newJavaId("com.exonum.binding:test-service-2:" + ARTIFACT_VERSION_2);
+      ServiceArtifactId.newJavaId("com.exonum.binding/test-service-2", ARTIFACT_VERSION_2);
   static final String SERVICE_NAME_2 = "test-service2";
   static final int SERVICE_ID_2 = 48;
 
@@ -53,13 +53,13 @@ final class TestKitTestUtils {
 
   static void createTestServiceArtifact(Path artifactsDirectory) throws IOException {
     Path artifactLocation = artifactsDirectory.resolve(ARTIFACT_FILENAME);
-    createArtifact(artifactLocation, ARTIFACT_ID, ARTIFACT_VERSION, TestServiceModule.class,
+    createArtifact(artifactLocation, ARTIFACT_ID, TestServiceModule.class,
         TestSchema.class, TestService.class);
   }
 
   static void createTestService2Artifact(Path artifactsDirectory) throws IOException {
     Path artifactLocation = artifactsDirectory.resolve(ARTIFACT_FILENAME_2);
-    createArtifact(artifactLocation, ARTIFACT_ID_2, ARTIFACT_VERSION_2, TestServiceModule2.class,
+    createArtifact(artifactLocation, ARTIFACT_ID_2, TestServiceModule2.class,
         TestSchema.class, TestService2.class);
   }
 
@@ -87,11 +87,10 @@ final class TestKitTestUtils {
   }
 
   private static void createArtifact(Path artifactLocation, ServiceArtifactId serviceArtifactId,
-                                     String serviceArtifactVersion, Class serviceModule,
-                                     Class<?>... artifactClasses) throws IOException {
+      Class serviceModule, Class<?>... artifactClasses) throws IOException {
     new ServiceArtifactBuilder()
-        .setPluginId(serviceArtifactId.getName())
-        .setPluginVersion(serviceArtifactVersion)
+        .setPluginId(serviceArtifactId.toString())
+        .setPluginVersion(serviceArtifactId.getVersion())
         .addClasses(artifactClasses)
         .addExtensionClass(serviceModule)
         .writeTo(artifactLocation);